+++ database/migrations/2024_12_03_104817_add_currency_id_packages.php
@@ -13,13 +13,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('packages', function (Blueprint $table) {
-            $table->foreignId('currency_id')->nullable()->constrained('global_currencies')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('packages', 'currency_id')) {
+            Schema::table('packages', function (Blueprint $table) {
+                $table->foreignId('currency_id')->nullable()->constrained('global_currencies')->onDelete('cascade');
+            });
+        }
 
-        Schema::table('restaurant_payments', function (Blueprint $table) {
-            $table->foreignId('package_id')->nullable()->constrained('packages')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('restaurant_payments', 'package_id')) {
+            Schema::table('restaurant_payments', function (Blueprint $table) {
+                $table->foreignId('package_id')->nullable()->constrained('packages')->onDelete('cascade');
+            });
+        }
 
         $defaultCurrency = GlobalCurrency::first();
         $defaultPackage = Package::first();
